Refactor app.py
* Move categories to helpers.py

* Move database models to models.py
