+++ models.py
@@ -1,0 +1,8 @@
+from app import db
+
+
+# Set up database tables
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(20), unique=True, nullable=False)
+    password = db.Column(db.String(150), nullable=False)
